# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Usuarios",
        "description": "API para la gestión completa de usuarios del sistema",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener lista de usuarios",
                "description": "Retorna una lista de todos los usuarios registrados en el sistema, ordenados alfabéticamente",
                "operationId": "7208c29037df2b8bb02499ba39ed8a78",
                "parameters": [
                    {
                        "name": "estado",
                        "in": "query",
                        "description": "Filtrar usuarios por estado",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "activo",
                                "inactivo"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de usuarios obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/User"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Crear un nuevo usuario",
                "description": "Registra un nuevo usuario en el sistema",
                "operationId": "262eb5ca1fdea354023c0a8870442c1e",
                "requestBody": {
                    "description": "Datos del usuario a crear",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/User"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "Usuario",
                "description": "Modelo de usuario del sistema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombres": {
                        "type": "string",
                        "example": "Juan Carlos"
                    },
                    "apellidos": {
                        "type": "string",
                        "example": "Pérez García"
                    },
                    "nombre_completo": {
                        "type": "string",
                        "example": "Juan Carlos Pérez García"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "juan.perez@example.com"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "+1234567890"
                    },
                    "estado": {
                        "type": "string",
                        "enum": [
                            "activo",
                            "inactivo"
                        ],
                        "example": "activo"
                    },
                    "fecha_registro": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15 10:30:00"
                    },
                    "fecha_ultima_modificacion": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-20 14:25:00"
                    }
                },
                "type": "object"
            },
            "UserRequest": {
                "title": "Solicitud de Usuario",
                "required": [
                    "nombres",
                    "apellidos",
                    "email",
                    "password"
                ],
                "properties": {
                    "nombres": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Juan Carlos"
                    },
                    "apellidos": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Pérez García"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 255,
                        "example": "juan.perez@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "minLength": 8,
                        "example": "password123"
                    },
                    "telefono": {
                        "type": "string",
                        "maxLength": 20,
                        "example": "+1234567890"
                    },
                    "estado": {
                        "type": "string",
                        "enum": [
                            "activo",
                            "inactivo"
                        ],
                        "example": "activo"
                    }
                },
                "type": "object"
            },
...
            "SuccessResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Operación exitosa"
                    },
                    "data": {
                        "type": "object"
                    },
                    "total": {
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error en la operación"
                    },
                    "error": {
                        "type": "string",
                        "example": "Mensaje de error detallado"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}